+++ Cargo.toml
@@ -1,3 +1,12 @@
-[workspace]
-members = ["database", "api_server", "migration", "entities"]
-resolver = "2"
+[package]
+name = "api_server"
+version = "0.0.1"
+edition = "2021"
+
+[dependencies]
+tokio = { version = "1.33.0", features = ["full"] }
+dotenv = "0.15.0"
+sea-orm = { version = "^0.12.0", features = [ "sqlx-postgres", "runtime-async-std-native-tls", "macros" ] }
+async-trait = { version = "0.1.73", features = [] }
+futures = "0.3.28"
+
