--- api_server/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "api_server"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-futures = "0.3.28"
-dotenv = "0.15.0"
-database = { path = "../database"}
-entities = { path = "../entities"}
-tokio = { version = "1.33.0", features = ["full"] }
