--- database/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "database"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-futures = "0.3.28"
-sea-orm = { version = "^0.12.0", features = [ "sqlx-postgres", "runtime-async-std-native-tls", "macros" ] }
-dotenv = "0.15.0"
-async-trait = "0.1.73"
-entities = { path = "../entities"}
