--- database/src/contexts/mod.rs
@@ -1,5 +1,0 @@
-pub mod DatabaseContext;
-pub mod EntityContext;
-pub mod ModelContext;
-pub mod SessionContext;
-pub mod UserContext;
