+++ src/entities/model.rs
