+++ src/entities/query.rs
