+++ src/entities/session.rs
